llm-v2/llmSecurityV1/requirements.txt
# llmSecurityV1 research framework dependencies
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pytest>=7.0.0
